backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
werkzeug
pyjwt
reportlab
gunicorn
psycopg2-binary
